Fix recursive dependency conflict
wget depends on openssl@1.1, so I had to update this formula to do the same.
Something about this feels wrong to me, but I poked around a little bit but
don't want to spend more time on it. Like why doesn't a specific version take
precedence over a non-version specified dependency? Also why is this an issue
if Homebrew can install both versions of openssl at the same time? It seems
it treats them as totally different entities in that the "@1.1" is part of
the name, rather than a version number that is meta data.